requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
